abstract spec
